requirements.txt
future
emoji
requests
sqlalchemy==1.3.20
python-telegram-bot==12.8
psycopg2-binary==2.8.6
feedparser
pyowm==2.10.0
googletrans
wikipedia
kbbi
Pillow
alphabet_detector
urbandict
coffeehouse
pynewtonmath
speedtest-cli
requests
wikipedia
tswift
tabulate
spongemock
geopy
Telethon==1.16.4
mtranslate
pylast
ffmpeg
youtube-search-python
psycopg2-binary==2.8.6
html2text
spotdl
langdetect==1.0.8
validators
pickledb
redis
pytz
pafy
gtts
bs4
lxml
demjson
wget
bs4
telegraph
hachoir
youtube-dl
bleach
alphabet_detector
tswift
covid
hurry.filesize
fontTools
bing_image_downloader
instantmusic
pybase64
pySmartDL
psutil
spamwatch
covid
markdown2>=2.3.10
zalgo_text
